cleanup

+++ board/safety/safety_honda.h
@@ -249,7 +249,7 @@ static void honda_rx_hook(const CANPacket_t *to_push) {
     }
   }
 
-  int bus_rdr_car = (honda_hw == HONDA_BOSCH) ? (int)0 : (int)2;  // radar bus, car side
+  int bus_rdr_car = (honda_hw == HONDA_BOSCH) ? 0 : 2;  // radar bus, car side
   bool stock_ecu_detected = false;
 
   // If steering controls messages are received on the destination bus, it's an indication
